reorder modules
