+++ starsim/__init__.py
@@ -22,30 +22,29 @@ sc.require(reqs, message=msg)
 del sc, reqs, msg # Don't keep this in the module
 
 # Finish imports
-from .utils            import *
-from .debug            import *
-from .arrays           import *
-from .time             import *
-from .parameters       import *
-from .distributions    import *
-from .people           import *
-from .modules          import *
-from .networks         import *
-from .results          import *
-from .demographics     import *
-from .products         import *
-from .interventions    import *
-from .analyzers        import *
-from .connectors       import *
-from .demographics     import *
-from .diseases         import *
-from .diseases         import *
-from .loop             import *
-from .sim              import *
-from .run              import *
-from .calibration      import *
-from .calib_components import *
-from .samples          import *
+from .utils         import *
+from .debugtools    import *
+from .arrays        import *
+from .time          import *
+from .parameters    import *
+from .distributions import *
+from .people        import *
+from .modules       import *
+from .networks      import *
+from .results       import *
+from .demographics  import *
+from .products      import *
+from .interventions import *
+from .analyzers     import *
+from .connectors    import *
+from .demographics  import *
+from .diseases      import *
+from .diseases      import *
+from .loop          import *
+from .sim           import *
+from .run           import *
+from .calibration   import *
+from .samples       import *
 
 # Load fonts
 def _load_fonts(debug=False):
