+++ starsim/debugtools.py
